App/requirements.txt
requests
splunk-opentelemetry[all]